requirements.txt
pdfplumber>=0.10
pdfminer.six>=20221105
PyMuPDF>=1.23
google-api-python-client>=2.100
google-auth-httplib2>=0.1
google-auth-oauthlib>=1.1
python-dotenv>=1.0
